requirements.txt
Flask==3.0.2
python-dotenv==1.0.1
flask-sqlalchemy==3.1.1
flask-migrate==4.0.5
flask-cors==4.0.0
pytest==8.0.2
black==24.2.0
flake8==7.0.0
PyJWT==2.8.0
email-validator==2.1.0.post1 